Rename mpgame.rs
