+++ src/main.rs
@@ -52,11 +52,11 @@ async fn main() {
             register(),
             commands::dict::lookup(),
             commands::dict::testmatch(),
-            commands::mpgame::worduel_challenge_timed(),
-            commands::mpgame::worduel_accept_timed(),
-            commands::mpgame::worduel_send_timed(),
-            commands::mpgame::worduel_forfeit_timed(),
-            commands::mpgame::worduel_keyboard_timed(),
+            commands::timedgame::worduel_challenge_timed(),
+            commands::timedgame::worduel_accept_timed(),
+            commands::timedgame::worduel_send_timed(),
+            commands::timedgame::worduel_forfeit_timed(),
+            commands::timedgame::worduel_keyboard_timed(),
             commands::misc::roll_dice(),
         ],
         prefix_options: poise::PrefixFrameworkOptions {
